requirements.txt
pycryptodome
PyJWT
Requests
beautifulsoup4
lxml
pytesseract
pillow